Use one permanent sandbox process per console session
This was changed to use a permanent process so the user has a consistent
expirience in the console, where all the commands are being run on the
same process.

Think on running `self()` several times. It now returns the same process
pid in each invokation (it was not the case with our previous
implementation).

It is also useful if the user want to play with `send` and `receive`
over the current process (relying on `self()`). However, we probably
won't support `send` and `receive` for security/integrity reasons, at
least in the first versions of the live console.