Run commands using tasks
Since we use an isolated process to run commands and we rely on `send`
and `receive` primitives, it is better to run each command using a `Task`.

This change is relevant because we were invoking the command execution
from a GenServer process (channel used by the live view connection) and
messages from our execution process were received alongside channel
messages. Mixing send/receive code with the existing messaging features
provided by LiveView processes seemed a bit risky.

If we decide to move the Sandbox code to a library, it is wise to do not
call `receive` in the caller process, so this move was going to happen
in any case.