Remove useless TODO comment
The comment was about the fact that the suggestions could contain the
same word that is going to be autocompleted. E.g. if we have `List`,
it would return `List` and do a void replacement of the word. Although,
it is doing unnecesary work, the code would be easier if we live with
this.