Add mechanism to validate functions invocations at runtime
There are a certain number of cases where it is impossible to detect
non-secure function invocations by only checking the code AST. For this
reason, we are adding a way to modify the code AST before execution.
With this new approach, an internal function of the Sandbox is going to
be executed prior to the user-code invocation, resulting in a runtime
error if the module or module/function to invocation is not included in
our whitelists.