Fix important security issue due to env passed to eval_quoted
In previous commits the full environment was passed to
Code.eval_quoted, but it made basic Erlang stuff accessible which is
very bad.

We must try to keep the env passed to eval_quoted as simple as possible
In fact, I'm now considering it could be a better approach to limit the
dangerous part of Elixir (further work needed to validate it).