Print bindings on sidebar when stored as tuples
In Elixir 1.10 the AST changed and the way bindings get stored can be
different in some cases.
In particular, when calling `eval_quoted` passing a quote and a list of
bindings, in Elixir 1.9 the current bindings were left intact if the
quoted expression didn't modify them, but in Elixir 1.10 they get
normalized.

Here's an example:

Elixir 1.9.4-otp-22
```
iex(1)> Code.eval_quoted(quote(do: Enum), [abc: 1], __ENV__)
{Enum, [abc: 1]}
```

Elixir 1.10.4-otp-22
```
iex(1)> Code.eval_quoted(quote(do: Enum), [abc: 1], __ENV__)
{Enum, [{{:abc, nil}, 1}]}
```

We can see that the bindings passed [abc: 1] are returned intact in the
first example, as the quote just returns the name of the Enum module.
However, in the second example, the bindings are normalized and the
first element of the tuple is the tuple {:abc, nil}.

The only error this was causing was when printing the bindings in the
sidebar component, a tuple can't be just be printed in HTML so when the
binding name is a tuple we only print the first part.
This could be reproduced before this commit, by first binding a variable
(abc = 1) and then evaluating a simple command like (nil).