Do not push a caret event on TAB keyup
This fixes a bug where two events, if received in a particular order
would position the caret incorrectly.
This was because the caret event is being handled on every keyUp event,
but when pressing the TAB key there's already the 'suggest' handler
updating the state, which also repositions the caret.
To fix this, we don't push the caret-position event anymore when hitting
TAB.