Stop syncing caret-position on the server
The value for the caret position is necessary only to suggest
autocompletions based on its position.
Now that we're able to tell where the caret is at the moment
that the user requests for an autocomplete, we don't need to keep
its position synced on the server on every key stroke, click or focus
events.
We still need to update it in the client when the input value changes in
the server after an autocomplete, or history cycle