Only reset history counter when executing a command
This helps reducing considerably the events sent through the socket as
we don't need to send messages anymore when a different key is pressed.

This changes the current behavior slightly.
Previously if I was cycling through the history and pressed any key,
even keys like shift or cmd, it would reset the history and using up
arrow after that would start from the begining.
Now we continue from the same position we were.